+++ next.config.mjs
@@ -117,7 +117,7 @@ const nextConfig = {
           },
           {
             key: "Permissions-Policy",
-            value: "camera=(), microphone=(), geolocation=(), browsing-topics=()",
+            value: "camera=(self), microphone=(self), geolocation=(), browsing-topics=()",
           },
           {
             key: "Content-Security-Policy",